+++ UI/SystemSettingsControl.xaml.cs
@@ -1,4 +1,5 @@
 using System.Windows.Controls;
+using System.Windows;
 
 namespace Hotkey_Translator.UI;
 
@@ -7,5 +8,12 @@ public partial class SystemSettingsControl : UserControl
     public SystemSettingsControl()
     {
         InitializeComponent();
+    }
+
+    public event RoutedEventHandler? ResetAllSettingsClicked;
+
+    private void OnResetAllSettingsClicked(object sender, RoutedEventArgs e)
+    {
+        ResetAllSettingsClicked?.Invoke(sender, e);
     }
 }